Voice Callback

+++ voice/response_voice.go
@@ -18,7 +18,7 @@ type Response struct {
 	PhoneNumberSid string
 	StartTime      common.JsonTime `json:"start_time"`
 	EndTime        common.JsonTime `json:"end_time"`
-	Duration       float64         `json:"duration"`
+	Duration       int             `json:"duration"`
 	AnsweredBy     string          `json:"answered_by"`
 	ForwardedFrom  string          `json:"fowarded_from"`
 	CallerName     string          `json:"caller_name"`
@@ -29,6 +29,7 @@ func (r *Response) Build(resp *http.Response) error {
 	if err != nil {
 		return errors.New(fmt.Sprintf("Error while reading json from buffer => %s", err.Error()))
 	}
+
 	err = json.Unmarshal(bodyBytes, r)
 	if err != nil {
 		return errors.New(fmt.Sprintf("Error while decoding json => %s, recieved msg => %s", err.Error(), string(bodyBytes)))
